[] (Tuan) optimize api

+++ src/main/java/com/finalmobile/wemarket/repository/DeliveryRepository.java
@@ -3,5 +3,5 @@ package com.finalmobile.wemarket.repository;
 import com.finalmobile.wemarket.models.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DeliveryRepository extends JpaRepository<Delivery,Long> {
+public interface DeliveryRepository extends JpaRepository<Delivery,Integer> {
 }
